api/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
alembic
python-dotenv
uvicorn
pydantic
requests
pytest
